refactor, import.sql issue

+++ src/main/java/ua/com/foxminded/serviceacc/controller/manager/ManagerController.java
@@ -27,18 +27,16 @@ import ua.com.foxminded.serviceacc.service.ManagerService;
 public class ManagerController implements Serializable {
 
 	private static final Logger log = LoggerFactory.getLogger(ManagerController.class);
-
 	private static final long serialVersionUID = 1L;
 
 	private Manager selectedManager;
     private List<ManagerInformation> managerInfo;
 
-	private List<Manager> managers;
 	private List<ManagerInformationType> managerInformationTypeList;
 
-	private ManagerService managerService;
-	private ManagerInformationTypeService managerInformationTypeService;
-	private ManagerInformationService managerInformationService;
+	private final ManagerService managerService;
+	private final ManagerInformationTypeService managerInformationTypeService;
+	private final ManagerInformationService managerInformationService;
 
 	@Inject
 	public ManagerController(ManagerService managerService,
@@ -51,7 +49,6 @@ public class ManagerController implements Serializable {
 
 	@PostConstruct
 	public void init() {
-		managers = managerService.findAll();
 		managerInformationTypeList = managerInformationTypeService.findAll();
 	}
 
@@ -60,69 +57,30 @@ public class ManagerController implements Serializable {
 		managerInfo = getManagerInformationList();
 	}
 
-	public void delete() {
-		managers.remove(selectedManager);
-		managerService.delete(selectedManager.getId());
-		selectedManager = null;
-	}
-
 	public void onOk() {
         //save or update manager
 	    if(selectedManager.getId() == null) {
             managerService.save(selectedManager);
-            managers.add(selectedManager);
         }else{
-            Manager updated = managerService.update(selectedManager);
-            int i = managers.indexOf(selectedManager);
-            managers.set(i, updated);
-            selectedManager = updated;
+            managerService.update(selectedManager);
         }
 
         //save or update information
         Iterator<ManagerInformation> iteratorInfos = managerInfo.iterator();
         while(iteratorInfos.hasNext()){
             ManagerInformation info = iteratorInfos.next();
-            if (info.getContent().isEmpty()) {
+            if (!info.getContent().isEmpty()) {
                 managerInformationService.update(info);
+            } else if (info.getContent().isEmpty() && info.getId() != null){
                 managerInformationService.delete(info.getId());
-            } else {
-                managerInformationService.update(info);
             }
         }
 
     }
 
-	public void setSelectedManager(Manager selectedManager) {
-		this.selectedManager = selectedManager;
-	}
+    public ManagerInformation getInfoByType(ManagerInformationType managerInformationType) {
 
-	public Manager getSelectedManager() {
-        return selectedManager;
-    }
-
-    public List<Manager> getList() {
-		return managers;
-	}
-
-	public ManagerInformationTypeService getManagerInformationTypeService() {
-		return managerInformationTypeService;
-	}
-
-	public void setManagerInformationTypeService(ManagerInformationTypeService managerInformationTypeService) {
-		this.managerInformationTypeService = managerInformationTypeService;
-	}
-
-	public List<ManagerInformationType> getManagerInformationTypeList() {
-		return managerInformationTypeList;
-	}
-
-	public void setManagerInformationTypeList(List<ManagerInformationType> managerInformationTypeList) {
-		this.managerInformationTypeList = managerInformationTypeList;
-	}
-
-	public ManagerInformation getInfoByType(ManagerInformationType managerInformationType) {
-
-	    if (selectedManager.getId() != null) {
+        if (selectedManager.getId() != null) {
             for (ManagerInformation managerInfo : managerInformationService.findByManager(selectedManager)) {
                 if (managerInfo.getManagerInformationType().equals(managerInformationType)) {
                     return managerInfo;
@@ -136,7 +94,7 @@ public class ManagerController implements Serializable {
         return managerInfo;
     }
 
-	public List<ManagerInformationType> getInfoTypeList(){ return managerInformationTypeService.findAll();}
+    public List<ManagerInformationType> getInfoTypeList(){ return managerInformationTypeList;}
 
     public List<ManagerInformation> getManagerInformationList() {
         managerInfo = new ArrayList<>();
@@ -147,4 +105,23 @@ public class ManagerController implements Serializable {
         return managerInfo;
     }
 
+	public void setSelectedManager(Manager selectedManager) {
+		this.selectedManager = selectedManager;
+	}
+
+	public Manager getSelectedManager() {
+        return selectedManager;
+    }
+
+	public ManagerInformationTypeService getManagerInformationTypeService() {
+		return managerInformationTypeService;
+	}
+
+	public List<ManagerInformationType> getManagerInformationTypeList() {
+		return managerInformationTypeList;
+	}
+
+	public void setManagerInformationTypeList(List<ManagerInformationType> managerInformationTypeList) {
+		this.managerInformationTypeList = managerInformationTypeList;
+	}
 }
